+++ Source/engine/render/text_render.hpp
@@ -6,12 +6,13 @@
 #pragma once
 
 #include <cstdint>
+#include <string>
 #include <utility>
 #include <vector>
 
 #include <SDL.h>
 
-#include "DiabloUI/ui_item.h"
+#include "DiabloUI/ui_flags.hpp"
 #include "engine.h"
 #include "engine/cel_sprite.hpp"
 #include "engine/rectangle.hpp"
