+++ Source/error.cpp
@@ -8,6 +8,7 @@
 
 #include "error.h"
 
+#include "DiabloUI/ui_flags.hpp"
 #include "control.h"
 #include "engine/render/cel_render.hpp"
 #include "engine/render/text_render.hpp"
