+++ Source/items.h
@@ -7,7 +7,7 @@
 
 #include <cstdint>
 
-#include "DiabloUI/ui_item.h"
+#include "DiabloUI/ui_flags.hpp"
 #include "engine.h"
 #include "engine/animationinfo.h"
 #include "engine/point.hpp"
