+++ Source/panels/charpanel.cpp
@@ -5,10 +5,10 @@
 
 #include "DiabloUI/art.h"
 #include "DiabloUI/art_draw.h"
-
 #include "control.h"
 #include "engine/render/cel_render.hpp"
 #include "engine/render/text_render.hpp"
+#include "panels/ui_panels.hpp"
 #include "player.h"
 #include "utils/display.h"
 #include "utils/language.h"
