+++ Source/panels/ui_panels.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+namespace devilution {
+
+enum class UiPanels {
+	Main,
+	Quest,
+	Character,
+	Spell,
+	Inventory,
+};
+
+} // namespace devilution
