+++ Source/utils/enum_traits.h
@@ -6,6 +6,7 @@
 #pragma once
 
 #include <cstddef>
+#include <type_traits>
 
 namespace devilution {
